requirements.txt
jinja2>=2.7
pyyaml
pymysql
pygithub
